Codecleanup from #581 (#583)
* is_visible_to_me checks now also in_visual_range
added obj::distance_to
fixed warnings

* no magic numbers for snprintf

+++ pol-core/pol/ssopt.h
@@ -45,7 +45,7 @@ struct ServSpecOpt
   unsigned int default_decay_time;
   bool refresh_decay_after_boat_moves;
   unsigned short default_doubleclick_range;
-  int default_accessible_range;
+  unsigned short default_accessible_range;
   unsigned short default_light_level;
   bool event_visibility_core_checks;
   unsigned short max_pathfind_range;
